05-Prompt-Engineering/requirements.txt
anthropic>=0.55
google-genai>=0.8
huggingface_hub>=0.27
openai>=1.40
pandas>=2.0
scikit-learn>=1.3
tenacity>=8.2
tqdm>=4.65
